Support for Responses API (#16)
NVIDIA NIM and NVIDIA Dynamo currently use the OpenAI ChatCompletions
API for text-to-text capabilities. With OpenAI’s introduction of the
Responses API — a stateful, event-driven surface that unifies truly
multimodal (image, text, audio) inputs and outputs, built-in tools, and
rich semantic SSE events — there is an opportunity for NVIDIA to define
a Responses API to serve as the http frontend, powered by the NIM/Dynamo
backend. This proposal details observations of several key aspects of
the OpenAI Responses API.

---------

Co-authored-by: nnshah1 <neelays@nvidia.com>