feat:  Container build process improvements to reduce build times and improve Dockerfile organization (#8)
This DEP introduces a container build process optimization strategy for
Dynamo to enhance developer experience and build efficiency by
reorganizing Dockerfiles. We are introducing a new Dynamo base container
which contains all the build logic to build Dynamo + NIXL wheels along
with an executable base container (with Dynamo + NIXL pre-installed)
that can be used for CI. We also define a structure for the
backend-specific Dockerfiles and introduce other build improvements such
as external caching.

---------

Signed-off-by: Tushar Sharma <tusharma@nvidia.com>